requirements.txt
Pillow
pyyaml
tqdm